+++ selene-lib/tests/lints/standard_library/ternary.lua
@@ -1,0 +1,2 @@
+math.floor(r > 1 and 1 or 0)
+math.floor(r > 1 + 1)