+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added `NumberRange` to the Roblox standard library.
 
 ## [0.4.3] - 2019-11-20
 ### Added