+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Unreleased
 ### Added
 - Added `NumberRange` to the Roblox standard library.
+- Added `string.split` to the Roblox standard library.
 
 ## [0.4.3] - 2019-11-20
 ### Added