+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Added `NumberRange` to the Roblox standard library.
 - Added `string.split` to the Roblox standard library.
 - Added `table.find` to the Roblox standard library.
+- Added `table.create` to the Roblox standard library.
+- Added `table.move` to the Roblox standard library.
+- Added `table.pack` to the Roblox standard library.
+- Added `table.unpack` to the Roblox standard library.
+- Added `coroutine.yieldable` to the Roblox standard library.
+- Added second argument to `math.log` to the Roblox standard library.
 
 ## [0.4.3] - 2019-11-20
 ### Added