+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Added `table.unpack` to the Roblox standard library.
 - Added `coroutine.yieldable` to the Roblox standard library.
 - Added second argument to `math.log` to the Roblox standard library.
+- Added `NumberSequenceKeypoint` to the Roblox standard library.
 
 ## [0.4.3] - 2019-11-20
 ### Added